refactor(GTFS+): move GTFS+ files around
