+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/Direction.java
@@ -1,16 +1,18 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class FareZoneAttribute extends Entity {
+public class Direction extends Entity {
 
     private static final long serialVersionUID = 1L;
 
-    public String zone_id;
-    public String zone_name;
+    public String route_id;
+    public int direction_id;
+    public String direction;
+
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
