+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/FareRiderCategory.java
@@ -1,16 +1,20 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.time.LocalDate;
 
-public class RealtimeTrip extends Entity {
+public class FareRiderCategory extends Entity {
 
     private static final long serialVersionUID = 1L;
 
-    public String trip_id;
-    public String realtime_trip_id;
+    public String fare_id;
+    public int rider_category_id;
+    public double price;
+    public LocalDate expiration_date;
+    public LocalDate commencement_date;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
