+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/FareZoneAttribute.java
@@ -1,18 +1,16 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class Direction extends Entity {
+public class FareZoneAttribute extends Entity {
 
     private static final long serialVersionUID = 1L;
 
-    public String route_id;
-    public int direction_id;
-    public String direction;
-
+    public String zone_id;
+    public String zone_name;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
