+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/RealtimeRoute.java
@@ -1,16 +1,17 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class TimePoint extends Entity {
-
+public class RealtimeRoute extends Entity {
     private static final long serialVersionUID = 1L;
 
-    public String trip_id;
-    public String stop_id;
+    public String route_id;
+    public int realtime_enabled;
+    public String realtime_routename;
+    public String realtime_routecode;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
