+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/RealtimeStop.java
@@ -1,19 +1,17 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class StopAttribute extends Entity {
+public class RealtimeStop extends Entity {
 
     private static final long serialVersionUID = 1L;
 
+    public String trip_id;
     public String stop_id;
-    public int accessibility_id;
-    public String cardinal_direction;
-    public String relative_position;
-    public String stop_city;
+    public String realtime_stop_id;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
