+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/StopAttribute.java
@@ -1,16 +1,19 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class RiderCategory extends Entity {
+public class StopAttribute extends Entity {
 
     private static final long serialVersionUID = 1L;
 
-    public int rider_category_id;
-    public String rider_category_description;
+    public String stop_id;
+    public int accessibility_id;
+    public String cardinal_direction;
+    public String relative_position;
+    public String stop_city;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
