+++ src/main/java/com/conveyal/datatools/manager/gtfsplus/tables/TimePoint.java
@@ -1,17 +1,16 @@
-package com.conveyal.datatools.manager.gtfsplus;
+package com.conveyal.datatools.manager.gtfsplus.tables;
 
 import com.conveyal.gtfs.model.Entity;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public class RealtimeStop extends Entity {
+public class TimePoint extends Entity {
 
     private static final long serialVersionUID = 1L;
 
     public String trip_id;
     public String stop_id;
-    public String realtime_stop_id;
 
     @Override
     public void setStatementParameters(PreparedStatement statement, boolean setDefaultId) throws SQLException {
